requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langchain-chroma
langchain-ollama
langchain-text-splitters
chromadb
pypdf
sentence-transformers